vulkan-hdr-layer-kwin6: init at v0.1
https://github.com/Zamundaaa/VK_hdr_layer

Enables HDR support for games (via "gamescope"), and programs like
"mpv", that use Vulkan's HDR extensions.

This is likely to be a temporary solution until Wayland upstream
adopts the necessary protocols, however that has been in the works for
many years:

https://gitlab.freedesktop.org/wayland/wayland-protocols/-/merge_requests/14

For the time being, this provides a path for users to experiment with
HDR, share feedback and hopefully make progress.

Arch provides a similar package:
https://aur.archlinux.org/packages/vk-hdr-layer-kwin6-git